backend/service-template/requirements.txt
# FastAPI и сервер
fastapi==0.110.0
uvicorn==0.29.0

# Валидация и сериализация
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1

# База данных
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0  # Для асинхронной работы с PostgreSQL

# Авторизация и аутентификация
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# Интеграция с другими сервисами
httpx==0.27.0  # Асинхронный HTTP-клиент
redis==5.0.2   # Клиент Redis для кеширования
pika==1.3.2    # Клиент RabbitMQ для очередей сообщений

# Тестирование
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.27.0

# Утилиты
python-dotenv==1.1.0
loguru==0.7.2   # Улучшенное логирование
tenacity==8.4.2 # Повторные попытки для функций 